refactor(cli): remove base app wrapper
