+++ bin/base/src/main.rs
@@ -5,17 +5,15 @@
 
 use clap::Parser;
 
-mod app;
 mod cli;
 mod config;
 
-use app::BaseApp;
 use cli::BaseCli;
 
 fn main() {
     base_cli_utils::init_common!();
 
-    if let Err(err) = BaseApp::new(BaseCli::parse()).run() {
+    if let Err(err) = BaseCli::parse().run() {
         eprintln!("Error: {err:?}");
         std::process::exit(1);
     }
